% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandnorm.R
\name{bandnorm}
\alias{bandnorm}
\title{BandNorm}
\usage{
bandnorm(path = NULL, hic_df = NULL, save = TRUE, save_path = NULL)
}
\arguments{
\item{path}{The path for all the cells in a directory. There can be sub-directories.}

\item{hic_df}{If you prepare the dataset as a format of "chrom", "binA", "binB", "count", "diag", "cell", you can input it into bandnorm directly.}

\item{save}{Whether to save each normalized cells. Default is TRUE. Note that if don't have large memory on your computer, and you need to use create_embedding function, it is highly recommended to save the cells because it helps lower the cost of memory in this function.}

\item{save_path}{Indicate the output path for normalized cells. Only need it when "save" parameter is TRUE. Default is NULL.}
}
\description{
This function allows you to calculate the BandNorm normalization.
}
\examples{
data("hic_df")
bandnorm_result = bandnorm(hic_df = hic_df, save = FALSE)
}
