% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandnorm.R
\name{bandnorm_cooler}
\alias{bandnorm_cooler}
\title{Cooler hic version of BandNorm (Thanks skelviper for supporting the code!)}
\usage{
bandnorm_cooler(coolerPath, resolution, cellname = NULL)
}
\description{
This function allows you to calculate the BandNorm normalization using the inputs that are cooler .mcool format. (We might also add .scool format.)
}
\examples{
# Will add example here.
}
