% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandnorm.R
\name{download_schic}
\alias{download_schic}
\title{Download existing single-cell raw data}
\usage{
download_schic(cell_line, cell_type = NULL, cell_path, summary_path = NULL)
}
\arguments{
\item{cell_line}{Must be one of "Kim2020", "Lee2019", "Li2019", "Ramani2017".}

\item{cell_type}{If you need to download a specific cell-type from one cell line, indicate the name of the cell-type in here.}

\item{cell_path}{Indicate the output path for raw data.}
}
\description{
This function allows you to calculate the BandNorm normalization.
}
\examples{
download_schic("Li2019", cell_path = getwd())
}
