% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandnorm.R
\name{plot_embedding}
\alias{plot_embedding}
\title{plot_embedding}
\usage{
plot_embedding(embedding, type, cell_type = NULL)
}
\arguments{
\item{embedding}{The PCA embedding obtained from create_embedding.}

\item{type}{A string of "tSNE" or "UMAP".}

\item{cell_type}{Include the cell type information in the plot. Default is NULL.}
}
\description{
This function allows you to calculate and plot the UMAP or tSNE embedding.
}
\examples{
bandnorm()
}
