% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGAD.R
\name{runProjection}
\alias{runProjection}
\title{Performing Projection for scGAD on other single-cell data}
\usage{
runProjection(DataList, doNorm, cellTypeList)
}
\arguments{
\item{DataList}{A list of matrices containing all assays. Each matrix should have a name that corresponds to the assay.}

\item{doNorm}{A vector of boolean. Each entries determines whether each matrix should be normalized.}
}
\description{
This function allows you to project scGAD value on various assays.
}
\examples{
#
}
