% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bandnorm.R
\name{plot_embedding}
\alias{plot_embedding}
\title{Plot tSNE or UMAP embedding}
\usage{
plot_embedding(embedding, type, cell_info = NULL, label = NULL)
}
\arguments{
\item{embedding}{The PCA embedding obtained from create_embedding.}

\item{type}{A string of "tSNE" or "UMAP".}

\item{cell_info}{Include the cell information in the plot. It should be a matrix or data.frame, and the first column is cell name in the corresponding to the embedding, and the second column is the information. Default is NULL.}

\item{label}{A string indicating the name of the cell information you want to include. Default is NULL.}
}
\description{
This function allows you to calculate and plot the UMAP or tSNE embedding.
}
\examples{
data("hic_df")
data("batch")
data("cell_type")
bandnorm_result = bandnorm(hic_df = hic_df, save = FALSE)
embedding = create_embedding(hic_df = bandnorm_result, do_harmony = TRUE, batch = batch)
plot_embedding(embedding, "UMAP", cell_info = cell_type, label = "Cell Type")
}
