% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGAD.R
\name{scGAD}
\alias{scGAD}
\title{Performing GAD for each single-cell data}
\usage{
scGAD(path, genes, cores = 10, output)
}
\arguments{
\item{path}{A path to the single-cell Hi-C data. The data format should be the same as bandnorm input.}

\item{genes}{A data frame containing 4 columns: chrmomsome, start, end, gene name.}

\item{cores}{Number of cores used for parallel running. Default is 10.}

\item{output}{The path for outputting scGAD. Each file contains two columns: gene name and gad value.}
}
\description{
This function allows you to calculate GAD value for each gene in each cell.
}
\examples{
#
}
