% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scGAD.R
\name{scGAD}
\alias{scGAD}
\title{Performing scGAD for each single-cell data}
\usage{
scGAD(
  path = NULL,
  hic_df = NULL,
  genes,
  depthNorm = TRUE,
  cores = 4,
  threads = 8
)
}
\arguments{
\item{path}{A path to the single-cell Hi-C data. The data format should be the same as bandnorm input.}

\item{hic_df}{You can also load dataframe containing all Hi-C data to here, but it is not recommended, since in scGAD, we are dealing with high resolution matrices, and it will consume a lot of memory if we load it directly.}

\item{genes}{A data frame containing 4 columns: chrmomsome, start, end, gene name.}

\item{depthNorm}{Whether to normalize the sequencing depth effect. Default is TRUE.}

\item{cores}{Number of cores used for parallel running. Default is 4.}

\item{threads}{Number of threads for fread function, default is 8.}
}
\description{
This function allows you to calculate GAD value for each gene in each cell.
}
\examples{
#
}
